Add edit office proper view
With initializing inputs with the old values

And make minor refactoring: add RecViewHelper and use it in implementations. Add keyboard visibility helper. Refactor that voivodeship dialog is from helper file. Show availability on open the dentist view. Add proper bottom paddings on edit and fix the visibility when keyboard is open.